+++ test/e2e/rules_api_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/prometheus/prometheus/rules"
 
 	"github.com/efficientgo/core/testutil"
-	"github.com/thanos-io/thanos/pkg/httpconfig"
+	"github.com/thanos-io/thanos/pkg/clientconfig"
 	"github.com/thanos-io/thanos/pkg/promclient"
 	"github.com/thanos-io/thanos/pkg/rules/rulespb"
 	"github.com/thanos-io/thanos/pkg/runutil"
@@ -67,11 +67,13 @@ func TestRulesAPI_Fanout(t *testing.T) {
 	)
 	testutil.Ok(t, e2e.StartAndWaitReady(prom1, sidecar1, prom2, sidecar2))
 
-	queryCfg := []httpconfig.Config{
+	queryCfg := []clientconfig.Config{
 		{
-			EndpointsConfig: httpconfig.EndpointsConfig{
-				StaticAddresses: []string{qBuilder.InternalEndpoint("http")},
-				Scheme:          "http",
+			HTTPConfig: clientconfig.HTTPConfig{
+				EndpointsConfig: clientconfig.HTTPEndpointsConfig{
+					StaticAddresses: []string{qBuilder.InternalEndpoint("http")},
+					Scheme:          "http",
+				},
 			},
 		},
 	}